+++ home/presentation/src/main/kotlin/com/rahim/yadino/home/presentation/model/RoutineUiModel.kt
@@ -14,6 +14,6 @@ data class RoutineUiModel(
   val id: Int? = null,
   val explanation: String? = null,
   val isSample: Boolean = false,
-  val idAlarm: Long? = null,
+  val idAlarm: Int? = null,
   val timeInMillisecond: Long? = null,
 )
