+++ routine/domain/src/main/kotlin/com/rahim/yadino/routine/domain/model/Routine.kt
@@ -12,6 +12,6 @@ data class Routine(
   val id: Int? = null,
   val explanation: String? = null,
   val isSample: Boolean = false,
-  val idAlarm: Long? = null,
+  val idAlarm: Int? = null,
   val timeInMillisecond: Long? = null,
 )
